+++ packages/angular-components/src/app/test/test.component.ts
@@ -1,8 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import {HtUsersService} from "../ht/ht-users.service";
-import {actionsClientFactory, dateRangeFactory} from "ht-client";
+import {actionsClientFactory, usersClientFactory, dateRangeFactory} from "ht-client";
 import {DateRangeMap} from "ht-data";
 import { actionClustersTrace } from "ht-maps"
+import {ActionsHeatmapTrace} from "ht-maps";
+import {HtMapService} from "../ht/ht-map.service";
 
 @Component({
   selector: 'ht-test',
@@ -14,17 +16,19 @@ export class TestComponent implements OnInit {
   userId = "43fbf0db-530b-4f79-9093-6f565ea6d37e";
   constructor(
     private userService: HtUsersService,
+    private mapService: HtMapService
   ) { }
 
   ngOnInit() {
     if (this.showAll) this.userService.setShowAll();
-    // let actionCluster = actionClustersTrace();
-    // let dateRangeService$ = dateRangeFactory(DateRangeMap.last_30_days);
-    // const client  = actionsClientFactory({dateRange$: dateRangeService$.data$.asObservable()});
+    // let actionCluster = new ActionsHeatmapTrace();
+    // let dateRangeService$ = dateRangeFactory(DateRangeMap.today);
+    // const client  = usersClientFactory({dateRange$: dateRangeService$.data$.asObservable()});
     // const list = client.list;
+    //
     // list.updateStrategy = 'once';
     // list.setActive(true);
-    // actionCluster.setData$(list.dataArray$)
+    // this.mapService.usersCluster.setData$(list.dataArray$)
 
   }