+++ packages/maps/src/entities/actions-heatmap.ts
@@ -5,7 +5,11 @@ import {DataObservableMixin, IMarkersArray, SetDataConfig} from "../mixins/data-
 import {Observable} from "rxjs/Observable";
 import {Subscription} from "rxjs/Subscription";
 import {StyleMixin} from "../mixins/styles";
-import {StyleFunct} from "../interfaces";
+import {Entity, StyleFunct} from "../interfaces";
+import {ExtendBoundsMixin} from "../mixins/extend-bounds";
+import {HtBounds} from "../map-utils/interfaces";
+import {MapInstance} from "../map-utils/map-instance";
+import {GlobalMap} from "../global/map-service";
 
 export class ActionsHeatmap {
   styleFunct: StyleFunct = {
@@ -30,13 +34,17 @@ export class ActionsHeatmap {
     }
   };
 
+  constructor(public mapInstance: MapInstance = GlobalMap) {}
+
   getPosition(item): HtPosition {
     return {
       lat: item.completed_place__location[1],
       lng: item.completed_place__location[0],
     }
-  }
+  };
 
 };
 
-export const ActionsHeatmapTrace = DataObservableMixin(HeatmapMixin(StyleMixin(ActionsHeatmap)));
+export const ActionsHeatmapTrace = ExtendBoundsMixin(
+  DataObservableMixin(ExtendBoundsMixin(HeatmapMixin(StyleMixin(ActionsHeatmap))))
+);