+++ packages/maps/src/mixins/heatmap.ts
@@ -1,11 +1,13 @@
 import { Constructor, Entities } from "../interfaces";
 import { HtPosition } from "ht-models";
-import { GlobalMap } from "../global/map-service";
 import {leafletHeat} from "../map-utils/leaflet.heatmap";
+import {HtBounds} from "../map-utils/interfaces";
+import {MapInstance} from "../map-utils/map-instance";
 
 export interface IHeatmapBase {
   getStyle: (styleType?) => object;
   getPosition: (data) => HtPosition;
+  mapInstance: MapInstance
 }
 
 export function HeatmapMixin<TBase extends Constructor<IHeatmapBase>>(
@@ -14,23 +16,25 @@ export function HeatmapMixin<TBase extends Constructor<IHeatmapBase>>(
   return class extends Base {
     map;
     heatmap;
-
+    forceExtendBounds = true;
+    entities: any[];
     constructor(...args: any[]) {
       super(...args);
-      let style = this.getStyle(GlobalMap.mapUtils.mapType);
-      this.heatmap = GlobalMap.mapUtils.getHeatmap(style);
+      let style = this.getStyle();
+      this.heatmap = this.mapInstance.mapUtils.getHeatmap(style);
     }
 
     trace(items: any[], map?) {
-      this.map = map || GlobalMap.map;
+      this.entities = items.map(item => ({item}));
+      this.map = map || this.mapInstance.map;
       if (this.map) {
         if(items) {
           let latLngs = items.reduce((acc, item) => {
             let position = this.getPosition(item);
             return position ? [...acc, position] : acc;
           }, []);
-          GlobalMap.mapUtils.updateHeatMapLatlng(latLngs, this.heatmap);
-          GlobalMap.mapUtils.setMap(this.heatmap, this.map)
+          this.mapInstance.mapUtils.updateHeatMapLatlng(latLngs, this.heatmap);
+          this.mapInstance.mapUtils.setMap(this.heatmap, this.map)
         } else {
           this.clear()
         }
@@ -38,6 +42,12 @@ export function HeatmapMixin<TBase extends Constructor<IHeatmapBase>>(
         console.warn("Map is not initialized");
         return false;
       }
+    }
+
+
+    getBounds(item, bounds?: HtBounds) {
+      let position = this.getPosition(item);
+      return this.mapInstance.mapUtils.extendBounds(position, bounds)
     }
 
     clear() {