+++ packages/maps/src/mixins/marker-renderer.ts
@@ -1,12 +1,14 @@
 import { Constructor, Entities } from "../interfaces";
-import { HtMap } from "../map-utils/interfaces";
+import {HtBounds, HtMap} from "../map-utils/interfaces";
 import { HtPosition } from "ht-models";
-import { GlobalMap } from "../global/map-service";
 import * as _ from "underscore";
+import {MapInstance} from "../map-utils/map-instance";
 
 export interface IMarkersBase {
   getStyle: (styleType?) => object;
   getPosition: (data) => HtPosition;
+  forceExtendBounds?: boolean;
+  mapInstance: MapInstance
 }
 
 export function MarkersMixin<TBase extends Constructor<IMarkersBase>>(
@@ -14,29 +16,25 @@ export function MarkersMixin<TBase extends Constructor<IMarkersBase>>(
 ) {
   return class extends Base {
     entities: Entities<any> = {};
-    cluster;
-
     htShow(item) {
       return `display: ${item ? "flex" : "none"}`;
     }
 
     getItem(data) {
-      return GlobalMap.mapUtils.getMarker();
+      return this.mapInstance.mapUtils.getMarker();
     }
 
-    getBounds(item, bounds?) {
-      return GlobalMap.mapUtils.extendBounds(item, bounds, !!this.cluster);
+    getBounds(item, bounds?): HtBounds {
+      return this.mapInstance.mapUtils.extendItemBounds(item, bounds, this.forceExtendBounds);
     }
 
     update({ item, data }) {
       let position = this.getPosition(data);
-      if (position) GlobalMap.mapUtils.updatePosition(item, position);
+      if (position) this.mapInstance.mapUtils.updatePosition(item, position);
     }
 
     removeItem(item) {
-      this.cluster &&
-        GlobalMap.mapUtils.removeClusterMarker(this.cluster, item);
-      GlobalMap.mapUtils.clearItem(item);
+      this.mapInstance.mapUtils.clearItem(item);
     }
 
     removeAll(entities) {
@@ -44,11 +42,6 @@ export function MarkersMixin<TBase extends Constructor<IMarkersBase>>(
         this.removeItem(entity.item);
       });
       this.entities = {};
-    }
-
-    setStyle(item) {
-      let style = this.getStyle();
-      GlobalMap.mapUtils.setStyle(item, style);
     }
 
     removeData(data) {