+++ packages/maps/src/mixins/styles.ts
@@ -1,9 +1,10 @@
-import { GlobalMap } from "../global/map-service";
 import { Constructor, StyleFunct } from "../interfaces";
+import {MapInstance} from "../map-utils/map-instance";
 
 export interface IStyleBase {
   styleFunct: StyleFunct;
   name?: string;
+  mapInstance: MapInstance
 }
 export function StyleMixin<TBase extends Constructor<IStyleBase>>(Base: TBase) {
   return class extends Base {
@@ -17,7 +18,8 @@ export function StyleMixin<TBase extends Constructor<IStyleBase>>(Base: TBase) {
     styleType: string;
 
     getStyle(selectedStyleType: string = "default", fallbackStyle?) {
-      const mapType = GlobalMap.mapUtils.type;
+      console.log(this.name);
+      const mapType = this.mapInstance.mapUtils.type;
       const styleFunct = this.styleFunct || this.defaultstyleFunct;
       const mapTypetype = styleFunct.get(mapType);
       // console.log(this.name, "style", selectedStyleType, styleFunct, this.styleFunct);
@@ -30,6 +32,12 @@ export function StyleMixin<TBase extends Constructor<IStyleBase>>(Base: TBase) {
           selectedStyleType
         );
       return style;
+    }
+
+
+    setStyle(item) {
+      let style = this.getStyle();
+      this.mapInstance.mapUtils.setStyle(item, style);
     }
   };
 }